+++ app/PCG/Computation/Internal.hs
@@ -1,3 +1,5 @@
+{-# LANGUAGE FlexibleContexts #-}
+
 module PCG.Computation.Internal
   ( evaluate
   , optimizeComputation
@@ -17,6 +19,7 @@ import qualified PCG.Command.Report.Evaluate as Report
 import qualified PCG.Command.Save.Evaluate   as Save
 import           PCG.Syntax
 import           System.Exit
+
 
 optimizeComputation :: Computation -> Computation
 optimizeComputation (Computation commands) = Computation $ collapseReadCommands commands