+++ src/Analysis/Parsimony/Dynamic/DirectOptimization/Pairwise/Internal.hs
@@ -44,21 +44,22 @@ module Analysis.Parsimony.Dynamic.DirectOptimization.Pairwise.Internal
   ) where
 
 import           Bio.Character.Encodable
-import           Control.Arrow            ((&&&))
+import           Control.Foldl              (Fold (..))
+import qualified Control.Foldl              as F
+import           Control.Monad.State.Strict
 import           Data.Bits
-import           Data.DList               (snoc)
+import           Data.DList                 (snoc)
 import           Data.Foldable
 import           Data.Key
-import           Data.List.NonEmpty       (NonEmpty (..))
-import qualified Data.List.NonEmpty       as NE
-import           Data.Matrix.NotStupid    (Matrix)
-import           Data.Maybe               (fromMaybe)
+import           Data.List.NonEmpty         (NonEmpty (..))
+import qualified Data.List.NonEmpty         as NE
+import           Data.Matrix.NotStupid      (Matrix)
+import           Data.Maybe                 (fromMaybe)
 import           Data.MonoTraversable
 import           Data.Ord
-import           Data.Semigroup
 import           Data.Semigroup.Foldable
 import           Numeric.Extended.Natural
-import           Prelude                  hiding (lookup)
+import           Prelude                    hiding (lookup)
 
 
 -- |
@@ -439,13 +440,52 @@ getOverlap inChar1 inChar2 costStruct = result
 -- the two (non-overlapping) least cost pairs are A,C and T,G, then the return
 -- value is A,C,G,T.
 overlap :: (EncodableStreamElement e {- , Show e -}) => (Word -> Word -> Word) -> e -> e -> (e, Word)
-overlap costStruct char1 char2
-  | intersectionStates == zero = minimalChoice $ symbolDistances costStruct char1 char2
-  | otherwise                  = (intersectionStates, 0)
+overlap costStruct char1 char2 = F.impurely ofoldMUnwrap (F.premapM g outerFold) char1 `evalState` 0
   where
-    intersectionStates = char1 .&. char2
-    zero = char1 `xor` char1
+    !zero     = char1 `xor` char1
 
+    outerFold = F.generalize $ Fold f (char1 `xor` char1, maxBound) id
+
+    costAndSymbol (i, x) = (x, cost1 + cost2)
+      where
+        !cost1 = getDistance3 costStruct i char1
+        !cost2 = getDistance3 costStruct i char2
+
+    f (!symbol1, !cost1) (!symbol2, !cost2) =
+        case cost1 `compare` cost2 of
+          EQ -> (symbol1 .|. symbol2, cost1)
+          LT -> (symbol1            , cost1)
+          GT -> (symbol2            , cost2)
+
+    g = const $ do
+        j <- get
+        modify' (+1)
+        let !v = toEnum j
+        let !b = zero `setBit` j
+        pure $ costAndSymbol (v, b)
+
+
+getDistance3 :: (MonoFoldable b, Element b ~ Bool) => (Word -> Word -> Word) -> Word -> b -> Word
+getDistance3 costStruct i b = fromMaybe errMsg $
+    F.impurely ofoldMUnwrap (F.premapM f (F.generalize F.minimum)) b `evalState` 0
+  where
+    errMsg = error "There were no bits set in the character!"
+
+    f e = do
+        j <- get
+        modify' (+1)
+        pure $ if e
+               then costStruct i j
+               else maxBound
+{-
+    getDistance2 :: FiniteBits b => Word -> b -> Word
+    getDistance2 i b =
+        case F.fold (F.prefilter (b `testBit`) (F.premap (costStruct i . toEnum) F.minimum)) indices of
+          Just x  -> x
+          Nothing -> error $ "There were no bits set in the character!"
+      where
+        indices = [0 .. finiteBitSize b - 1]
+-}
 
 -- |
 -- Given a structure of unambiguous character elements and costs, calculates the
@@ -461,6 +501,7 @@ minimalChoice = foldl1 f
           GT -> (symbol2            , cost2)
 
 
+{-
 -- |
 -- Finds the cost between all single, unambiguous symbols and two dynamic
 -- character elements (ambiguity groups of symbols).
@@ -479,24 +520,55 @@ symbolDistances costStruct char1 char2 = costAndSymbol <$> allSymbols
   where
     costAndSymbol (i, x) = (x, cost1 + cost2)
       where
-        cost1 = getDistance i char1
-        cost2 = getDistance i char2
+        !cost1 = getDistance3 costStruct i char1
+        !cost2 = getDistance3 costStruct i char2
 
     symbolIndices = NE.fromList [0 .. finiteBitSize char1 - 1]
     allSymbols    = (toEnum &&& setBit zero) <$> symbolIndices
     zero          = char1 `xor` char1
+-}
 
+{-
+    getDistance3 :: (MonoFoldable b, Element b ~ Bool) => Word -> b -> Word
+    getDistance3 i b =
+        case F.impurely ofoldMUnwrap (F.prefilterM pure (F.premapM f (F.generalize F.minimum))) b `evalState` 0 of
+          Just x  -> x
+          Nothing -> error $ "There were no bits set in the character!"
+      where
+        f _ = do
+            j <- get
+            modify' (+1)
+            pure $ costStruct i j
+-}
+
+{-
+    getDistance2 :: FiniteBits b => Word -> b -> Word
+    getDistance2 i b =
+        case F.fold (F.prefilter (b `testBit`) (F.premap (costStruct i . toEnum) F.minimum)) indices of
+          Just x  -> x
+          Nothing -> error $ "There were no bits set in the character!"
+      where
+        indices = [0 .. finiteBitSize b - 1]
+-}
+{-
     getDistance :: FiniteBits b => Word -> b -> Word
-    getDistance i e = minimum $ costStruct i <$> getSetBits e
+    getDistance i e = F.minimum $ costStruct i <$> getSetBits e
 
     getSetBits :: FiniteBits b => b -> NonEmpty Word
     getSetBits b =
-        case filter (b `testBit`) indices of
-          x:xs -> toEnum <$> x:|xs
+        case fold (F.prefilter (b `testBit`) (costStruct i <$> F.minimum)) indices of
+          x:xs -> x:|xs
           []   -> error $ "There were no bits set in the character: " <>
                     show (foldMap (\i -> if b `testBit` i then "1" else "0") indices)
       where
         indices = [0 .. finiteBitSize b - 1]
+-}
+{-
+        go  0 xs = if b `testBit` 0 then 0:xs else xs
+        go !n xs
+          | b `testBit` n = go (n-1) $ toEnum n:xs
+          | otherwise     = go (n-1) xs
+-}
 
 
 -- |