+++ src/Test/Custom/DynamicCharacterNode.hs
@@ -74,10 +74,13 @@ constructNode lhs rhs
     rootDec = toRootNode lhsDec rhsDec
 
 
-toLeafNode :: ( Ord (Element c)
-              , SimpleDynamicDecoration d c
-              )
-           => d -> DynamicDecorationDirectOptimizationPostorderResult c
+toLeafNode
+  :: ( Ord (Element c)
+     , SimpleDynamicDecoration d c
+--     , GetSparseTransitionCostMatrix (DynamicCharacterMetadataDec (Element c)) MemoizedCostMatrix
+     )
+  => d
+  -> DynamicDecorationDirectOptimizationPostorderResult c
 toLeafNode c = directOptimizationPostorder pairwiseFunction (LeafContext c)
 
 
@@ -105,7 +108,7 @@ scm i j = if i == j then 0 else 1
 
 
 defMetadata :: DynamicCharacterMetadataDec (Element DynamicCharacter)
-defMetadata = dynamicMetadata defName defWeight defAlphabet scm Nothing
+defMetadata = dynamicMetadataWithTCM defName defWeight defAlphabet scm
 
 
 initDec :: DynamicCharacter -> DynamicDecorationInitial DynamicCharacter
@@ -122,3 +125,4 @@ defWeight = 1
 
 defAlphabet :: Alphabet String
 defAlphabet = fromSymbols ["A","C","G","T"]
+