+++ test/TestSuite/ScriptTests.hs
@@ -1,5 +1,6 @@
 {-# LANGUAGE FlexibleContexts  #-}
 {-# LANGUAGE OverloadedStrings #-}
+{-# LANGUAGE Strict            #-}
 {-# LANGUAGE TypeFamilies      #-}
 
 module TestSuite.ScriptTests
@@ -30,26 +31,28 @@ import           Test.Tasty.HUnit
 import           Text.Megaparsec
 import           Text.Megaparsec.Char
 import           Text.Megaparsec.Char.Lexer        (scientific)
-import           Turtle                     hiding (char, many, satisfy, x)
-import qualified Turtle                     as T
 
 
 import Control.Arrow              ((&&&))
 import Control.DeepSeq
+import Data.Bifunctor             (first)
 import Data.Char                  (isSpace)
 import Data.Either
-import Data.Foldable
-import Data.Functor               (($>))
+import Data.Functor               (void, ($>))
 import Data.Scientific            hiding (scientific)
 import Data.Text                  (Text)
+import Data.Text.IO               (readFile)
 import Data.Void                  (Void)
 import Numeric.Extended.Real
+import Prelude                    hiding (readFile, writeFile)
+import System.Exit
+import System.Process
 import Test.Tasty
 import Test.Tasty.HUnit
+import TestSuite.SubProcess
 import Text.Megaparsec
 import Text.Megaparsec.Char
 import Text.Megaparsec.Char.Lexer (scientific)
-import Turtle                     hiding (char, many, parallel, satisfy, wait, x)
 
 
 data  ExecutionResults
@@ -59,27 +62,24 @@ data  ExecutionResults
     }
 
 
-testSuite :: IO TestTree
-testSuite = testGroup "Script Test Suite" <$> sequenceA
+testSuite :: TestTree
+testSuite = testGroup "Script Test Suite"
   [ scriptCheckCost 50.46
         "datasets/continuous/single-block/arthropods.pcg"
         "datasets/continuous/single-block/cost.data"
-  , scriptCheckValue getCost 50.46
-        "datasets/continuous/single-block/arthropods.pcg"
+  ,  scriptCheckCost 7
+        "datasets/continuous/missing/test.pcg"
+        "datasets/continuous/missing/cost.data"
+  , scriptCheckCost 8
+        "datasets/non-additive/missing/test.pcg"
+        "datasets/non-additive/missing/cost.data"
   , scriptCheckCost 1665
         "datasets/non-additive/single-block/arthropods.pcg"
         "datasets/non-additive/single-block/cost.data"
-    -- The missing tests are commented out as they are not curretnly on master
---  ,  scriptCheckCost 7
---        "datasets/continuous/missing/test.pcg"
---        "datasets/continuous/missing/cost.data"
---  , scriptCheckCost 8
---        "datasets/non-additive/missing/test.pcg"
---        "datasets/non-additive/missing/cost.data"
---  , scriptCheckCost 1206.0
---        "datasets/additive/missing/test.pcg"
---        "datasets/additive/missing/cost.data"
-  , scriptCheckCost 77252
+  , scriptCheckCost 56
+        "datasets/additive/missing/test.pcg"
+        "datasets/additive/missing/cost.data"
+  , scriptCheckCost 2642
         "datasets/additive/single-block/arthropods.pcg"
         "datasets/additive/single-block/cost.data"
   , scriptCheckCost 1
@@ -94,6 +94,12 @@ testSuite = testGroup "Script Test Suite" <$> sequenceA
   , scriptCheckCost 4
         "datasets/additive/case-3/case-3.pcg"
         "datasets/additive/case-3/cost.data"
+  , scriptCheckCost 16
+      "datasets/sankoff/multi-block/missing/missing.pcg"
+      "datasets/sankoff/multi-block/missing/cost.data"
+  , scriptCheckCost 12
+      "datasets/sankoff/single-block/missing/missing-values.pcg"
+      "datasets/sankoff/single-block/missing/cost.data"
   , scriptCheckCost 914
         "datasets/sankoff/single-block/dna/discrete/arthropods.pcg"
         "datasets/sankoff/single-block/dna/discrete/cost.data"
@@ -109,7 +115,7 @@ testSuite = testGroup "Script Test Suite" <$> sequenceA
   , scriptCheckCost 1143
         "datasets/sankoff/single-block/protein/discrete/invertebrates.pcg"
         "datasets/sankoff/single-block/protein/discrete/cost.data"
-  , scriptCheckCost 11851
+  , scriptCheckCost 11813
         "datasets/sankoff/single-block/protein/L1-norm/invertebrates.pcg"
         "datasets/sankoff/single-block/protein/L1-norm/cost.data"
   , scriptCheckCost 2012
@@ -172,34 +178,33 @@ testSuite = testGroup "Script Test Suite" <$> sequenceA
   , scriptCheckCost 230
         "datasets/sankoff/single-block/huge-mix/levenshtein/test.pcg"
         "datasets/sankoff/single-block/huge-mix/levenshtein/cost.data"
+  , scriptCheckCost 45
+      "datasets/dynamic/multi-block/missing/missing.pcg"
+      "datasets/dynamic/multi-block/missing/cost.data"
   , scriptCheckCost 2042
-        "datasets/dynamic/multi-block/arthropods.pcg"
-        "datasets/dynamic/multi-block/cost.data"
+      "datasets/dynamic/multi-block/dna/arthropods.pcg"
+      "datasets/dynamic/multi-block/dna/cost.data"
+  , scriptCheckCost 28
+      "datasets/dynamic/single-block/missing/missing-values.pcg"
+      "datasets/dynamic/single-block/missing/cost.data"
+  , scriptCheckCost 11036
+      "datasets/dynamic/single-block/protein/L1-norm/invertebrates.pcg"
+      "datasets/dynamic/single-block/protein/L1-norm/cost.data"
   , scriptCheckCost 1132
-        "datasets/dynamic/single-block/protein/discrete/invertebrates.pcg"
-        "datasets/dynamic/single-block/protein/discrete/cost.data"
-
-  , scriptCheckCost 2042
-        "datasets/dynamic/single-block/protein/L1-norm/invertebrates.pcg"
-        "datasets/dynamic/single-block/protein/L1-norm/cost.data"
-
-{--
+      "datasets/dynamic/single-block/protein/discrete/invertebrates.pcg"
+      "datasets/dynamic/single-block/protein/discrete/cost.data"
   , scriptCheckCost 1948
         "datasets/dynamic/single-block/protein/1-2/invertebrates.pcg"
         "datasets/dynamic/single-block/protein/1-2/cost.data"
   , scriptCheckCost 1241
         "datasets/dynamic/single-block/protein/2-1/invertebrates.pcg"
         "datasets/dynamic/single-block/protein/2-1/cost.data"
+  , scriptCheckCost 3413
+      "datasets/dynamic/single-block/slashes/L1-norm/test.pcg"
+      "datasets/dynamic/single-block/slashes/L1-norm/cost.data"
   , scriptCheckCost 197
         "datasets/dynamic/single-block/slashes/discrete/test.pcg"
         "datasets/dynamic/single-block/slashes/discrete/cost.data"
---}
-
-  , scriptCheckCost 2042
-        "datasets/dynamic/single-block/slashes/L1-norm/test.pcg"
-        "datasets/dynamic/single-block/slashes/L1-norm/cost.data"
-
-{--
   , scriptCheckCost 254
         "datasets/dynamic/single-block/slashes/1-2/test.pcg"
         "datasets/dynamic/single-block/slashes/1-2/cost.data"
@@ -212,43 +217,42 @@ testSuite = testGroup "Script Test Suite" <$> sequenceA
   , scriptCheckCost 488
         "datasets/dynamic/single-block/slashes/levenshtein/test.pcg"
         "datasets/dynamic/single-block/slashes/levenshtein/cost.data"
-  , scriptCheckCost 197
+  , scriptCheckCost 133
         "datasets/dynamic/single-block/large-mix/discrete/test.pcg"
         "datasets/dynamic/single-block/large-mix/discrete/cost.data"
-  , scriptCheckCost 2042
+  , scriptCheckCost 7185
         "datasets/dynamic/single-block/large-mix/L1-norm/test.pcg"
         "datasets/dynamic/single-block/large-mix/L1-norm/cost.data"
-  , scriptCheckCost 254
+  , scriptCheckCost 164
         "datasets/dynamic/single-block/large-mix/1-2/test.pcg"
         "datasets/dynamic/single-block/large-mix/1-2/cost.data"
-  , scriptCheckCost 228
+  , scriptCheckCost 172
         "datasets/dynamic/single-block/large-mix/2-1/test.pcg"
         "datasets/dynamic/single-block/large-mix/2-1/cost.data"
-  , scriptCheckCost 671
+  , scriptCheckCost 367
         "datasets/dynamic/single-block/large-mix/hamming/test.pcg"
         "datasets/dynamic/single-block/large-mix/hamming/cost.data"
-  , scriptCheckCost 488
+  , scriptCheckCost 213
         "datasets/dynamic/single-block/large-mix/levenshtein/test.pcg"
         "datasets/dynamic/single-block/large-mix/levenshtein/cost.data"
-  , scriptCheckCost 197
+  , scriptCheckCost 246
         "datasets/dynamic/single-block/huge-mix/discrete/test.pcg"
         "datasets/dynamic/single-block/huge-mix/discrete/cost.data"
-  , scriptCheckCost 2042
+  , scriptCheckCost 21753
         "datasets/dynamic/single-block/huge-mix/L1-norm/test.pcg"
         "datasets/dynamic/single-block/huge-mix/L1-norm/cost.data"
-  , scriptCheckCost 254
+  , scriptCheckCost 325
         "datasets/dynamic/single-block/huge-mix/1-2/test.pcg"
         "datasets/dynamic/single-block/huge-mix/1-2/cost.data"
-  , scriptCheckCost 228
+  , scriptCheckCost 284
         "datasets/dynamic/single-block/huge-mix/2-1/test.pcg"
         "datasets/dynamic/single-block/huge-mix/2-1/cost.data"
-  , scriptCheckCost 671
+  , scriptCheckCost 872
         "datasets/dynamic/single-block/huge-mix/hamming/test.pcg"
         "datasets/dynamic/single-block/huge-mix/hamming/cost.data"
-  , scriptCheckCost 488
+  , scriptCheckCost 698
         "datasets/dynamic/single-block/huge-mix/levenshtein/test.pcg"
         "datasets/dynamic/single-block/huge-mix/levenshtein/cost.data"
---}
   , scriptFailure "datasets/unmatched-leaf-taxon/test.pcg"
   , scriptFailure "datasets/unmatched-tree-taxon/test.pcg"
   , scriptFailure "datasets/duplicate-leaf-taxon/test.pcg"
@@ -311,10 +315,11 @@ getCost = either
 parseCost :: Text -> Maybe ExtendedReal
 parseCost = parseMaybe fileSpec
   where
+    parseResult :: Either (ParseErrorBundle Text Void) ExtendedReal
+    parseResult = parse fileSpec path $ force str
+
     fileSpec =  many (try (ignoredLine <* notFollowedBy costLine))
-             *> ignoredLine
-             *> costLine
-             <* many ignoredLine
+             *> (try costLine <|> (ignoredLine *> costLine))
 
     costLine :: MonadParsec Void Text m => m ExtendedReal
     costLine =  many inlineSpace
@@ -323,12 +328,11 @@ parseCost = parseMaybe fileSpec
              *> extendedReal
              <* ignoredLine
 
-
     extendedReal :: MonadParsec Void Text m => m ExtendedReal
     extendedReal = (fromFinite . toRealFloat <$> scientific) <|> (char '∞' $> infinity)
 
     ignoredLine  :: MonadParsec Void Text m => m ()
-    ignoredLine  = many inlineChar *> newlineChar
+    ignoredLine  = many inlineChar *> newlineChar -- (newlineChar <|> eof)
 
     inlineSpace  :: MonadParsec Void Text m => m ()
     inlineSpace  = void $ satisfy (\x -> isSpace x && x /= '\n')
@@ -340,9 +344,10 @@ parseCost = parseMaybe fileSpec
     newlineChar  = void $ char '\n'
 
 
-
-scriptFailure :: String -> IO TestTree
-scriptFailure scriptPath = scriptTest scriptPath [] (testCase scriptPath . assertBool "Expected script failure" . isLeft)
+scriptFailure :: String -> TestTree
+scriptFailure scriptPath = testCase scriptPath $ do
+    v <- runExecutable scriptPath []
+    assertBool "Expected script failure" $ isLeft v
 
 
 -- |
@@ -357,26 +362,106 @@ runExecutable
                                       --   graph state at the end of the script (if applicable) and files,
                                       --   or the exit code if the script failed
 runExecutable scriptStr outputPaths = do
-    startingDirectory <- pwd
-    cd scriptDirectory
-    (exitCode, _) <- shellStrict ("stack exec pcg -- --input " <> scriptText <> " --output test.log") mempty
-    cd startingDirectory
+--
+{-
+    dbFile <- openFile "output.debug" WriteMode
+    hSetBuffering dbFile NoBuffering
+    hPutStrLn dbFile startingDirectory
+    hPutStrLn dbFile scriptDirectory
+    hPutStrLn dbFile absScriptDirectory
+    _      <- hClose dbFile
+-}
+    ctx <- constructProcess scriptStr
+    (exitCode, _, _) <- readCreateProcessWithExitCode (process ctx) mempty
+--                            (defineProcess absScriptDirectory command) mempty
+--                            (defineProcess scriptDirectory command) mempty
+    _ <- destructProcess ctx
     case exitCode of
-      ExitFailure v -> cd startingDirectory $> Left v
-      _             -> do
-          foundFile <- doesFileExist defaultSaveFilePath
-          saveState <- if   not foundFile
-                       then pure $ Left "File not found :("
-                       else fmap getCompact
-                         <$> unsafeReadCompact defaultSaveFilePath :: IO (Either String GraphState)
-          cd startingDirectory
-          Right . ExeRes saveState <$> traverse (readTextFile . decodeString) outputPaths
+      ExitFailure v -> pure  $ Left v
+      _             -> force . Right <$> traverse readFile outputPaths
+--  where
+--    command :: String
+--    command = "stack exec pcg -- --input " <> scriptFile <> " --output test.log"
 
+
+{-
+outLogFileName :: String
+outLogFileName = "log.out"
+
+
+errLogFileName :: String
+errLogFileName = "log.err"
+
+
+data  ScriptContext
+    = ScriptContext
+    { process :: CreateProcess
+    , outPath :: FilePath
+    , errPath :: FilePath
+    }
+
+
+constructProcess
+  :: FilePath -- ^ Relative path to the PCG script
+  -> IO ScriptContext
+constructProcess scriptStr = do
+    prefix <- makeAbsolute scriptDirectory
+    let outFilePath = prefix </> outLogFileName
+    let errFilePath = prefix </> errLogFileName
+    let commandStr  = unwords
+                    [ "stack exec pcg -- --input"
+                    , scriptFileName
+                    , "--output"
+                    , outFilePath
+                    , "2>"
+                    , errFilePath
+                    ]
+
+    let p = CreateProcess
+            { cmdspec            = ShellCommand commandStr
+            , cwd                = Just scriptDirectory
+            , env                = Nothing
+            -- Do not use the stream handles, they do not work with Tasty
+            , std_in             = NoStream
+            , std_out            = NoStream
+            , std_err            = NoStream
+            , close_fds          = True
+            , create_group       = False
+            , delegate_ctlc      = False
+            , detach_console     = False
+            , create_new_console = False
+            , new_session        = False
+            , child_group        = Nothing
+            , child_user         = Nothing
+            , use_process_jobs   = False
+            }
+
+    pure ScriptContext
+        { process = p
+        , outPath = outFilePath
+        , errPath = errFilePath
+        }
   where
-    scriptText = either id id $ toText scriptFile
+    (scriptDirectory, scriptFileName) = breakScriptPath scriptStr
 
-    (scriptDirectory, scriptFile) = breakScriptPath $ decodeString scriptStr
-
-    breakScriptPath = (collapse . foldl' (</>) defaultDirectory . init &&& last) . splitDirectories
+    breakScriptPath = (normalise . foldl' (</>) defaultDirectory . init &&& last) . splitDirectories
       where
-        defaultDirectory = decodeString "."
+        defaultDirectory = "."
+
+
+destructProcess :: ScriptContext -> IO ()
+destructProcess ctx = mapM_ cleanUpHandle $
+    [ (std_out . process &&& outPath)
+    , (std_err . process &&& errPath)
+    ] <*> [ ctx ]
+  where
+    cleanUpHandle (s, p) = do
+        -- If we have an opened handle, close it
+        _ <- case s of
+               UseHandle h -> hClose h
+               _           -> pure ()
+        n <- getFileSize p
+        if n == 0 -- If no data was written, delete the file
+        then removeFile p
+        else pure ()
+-}