+++ CHANGELOG.md
@@ -10,10 +10,12 @@ Releases are available [on GitHub][2].
 * Using semantic versioning
 * Updated to GHC-8.4.4
 * Updated to Cabal-2.2
+* Updated build command to add network edges to input graphs rather than wagner build a new tree first.
 * Updated read command grammar to allow for more specific TCM specification
 * Improved efficiency and consistency of character metadata representation
 * Improved efficiency of Wagner build
 * Improved efficiency of Data.Vector.NonEmpty.fromNonEmpty, unfoldr, and traverse
+* Improved efficiency of three-way median and cost calculation (discrete metric, L1 norm, and small alphabet specialization)
 * Improved representation of character & metadata sequences
 * Improved representation of Symbol Change Matrix (SCM) and Transition Cost Matrix (TCM)
 * Improved representation of the types of postorder and preorder functions
@@ -30,11 +32,16 @@ Releases are available [on GitHub][2].
 * Added module for stricter fold functions
 * Added lenses for many graph and node accessors
 * Added makefile commands for outputting core, building with llvm backend, various test options and building a local hoogle database
+* Added module for memoized vector computation abstracting the method used in pre/postorder
 * Streamlined build experience on MacOS
 * Corrected defect in handling of prealigned input data
+* Corrected defect in FFI resulting in a spae leak
+* Corrected defect in Data.Range value construction
+* Corrected defect in Data.Range.Ranged type-class definitions
 * Corrected defect in Data.Vector.NonEmpty.unfoldr
 * Corrected defect in headEx & lastEx of several MonoFoldable instances
 * Corrected defect in omap over missing dynamic characters
+* Corrected defect in candidateNetworkEdges to not return inconsistent edges
 * Streamlined build experience on MacOS
 * Polished command line option parsing and error reporting
 * Improved phylogeny pretty-printing